Give feedback when using switches in listings.
Switch to using OnClick instead of OnCheckedChanged, since the latter
was called by updating the data binding, whereas we only want to
communicate with the server when the user performs the switch.

ServerException was created to properly handle error messages retrieved
from the server, and are appropriately added to the HttpStatusException
as a formatted String.